# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Radio call manager API — real-time audio streaming, call management, and admin CRUD.",
        "title": "OpenScanner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/admin/shared-links": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all shared links with associated call metadata.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List shared links",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.SharedLinkListItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/shared-links/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a shared link by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a shared link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shared link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted: true",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/tools/audio-missing": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginates through stored calls and returns those whose audio file is missing from disk.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List calls with missing audio files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page size (1–1000, default 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.missingAudioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/tools/audio-missing/cleanup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes call database rows for the given call IDs, but only if their audio file is still missing on disk at delete time. Requires confirm=true.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Clean up calls with missing audio",
                "parameters": [
                    {
                        "description": "Call IDs to clean up",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.cleanupMissingAudioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.cleanupMissingAudioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "descriptive error message"
                }
            }
        },
...
        "internal_api.SharedLinkListItem": {
            "type": "object",
            "properties": {
                "callId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "integer"
                },
                "dateTime": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "sharedBy": {
                    "type": "string"
                },
                "systemLabel": {
                    "type": "string"
                },
                "talkgroupLabel": {
                    "type": "string"
                },
                "talkgroupName": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.cleanupMissingAudioRequest": {
            "type": "object",
            "properties": {
                "callIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "confirm": {
                    "type": "boolean"
                }
            }
        },
        "internal_api.cleanupMissingAudioResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "integer"
                },
                "requested": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.missingAudioCall"
                    }
                }
            }
        },
...
        "internal_api.missingAudioCall": {
            "type": "object",
            "properties": {
                "audioName": {
                    "type": "string"
                },
                "audioPath": {
                    "type": "string"
                },
                "dateTime": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "internal_api.missingAudioResponse": {
            "type": "object",
            "properties": {
                "checked": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "missing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.missingAudioCall"
                    }
                },
                "offset": {
                    "type": "integer"
                },
                "recordingsDir": {
                    "type": "string"
                },
                "totalCalls": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "API key for call upload endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token prefixed with \"Bearer \"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}